sifive_prci: Emulate the plloutdiv register
This register contains the post-PLL output divider, which divides the
PLL's output again to produce the final PLL output clock speed.

QEMU used to abort when reading from this register, but it's a valid
register to read from.  We noticed the problem when improving the clock
driver support for the HiFive1 in embedded land.

This patch emulates reads from this register by providing the reset
value from the HiFive1 and ignoring writes (like we do with the other
PRCI registers).

I haven't even compiled this patch.

Signed-off-by: Palmer Dabbelt <palmer@sifive.com>