sifive_prci: Emulate the HFXOSC configuration register
I figured that when I went to touch the PRCI I might as well read the
documentation and make it match the HiFive1.  This patch adds the
external oscillator control register, which reads as the most likely
value (it's enabled on reset, and it's probably locked by the time
you're trying to read it).

Like the others, I haven't compiled this.

Signed-off-by: Palmer Dabbelt <palmer@sifive.com>