sifive_prci: Mark the HFROSC as enabled
The HRFOSC has a two-bit handshake mechanism: software sets the enabled
bit, and then checks for the ready bit to ensure the oscillator has
locked.  The value we used to return had the ready bit set but the
enabled bit clear, which is a non-sensical value: the oscillator won't
be ready if it's off.

This patch changes the value returned for reads of the HFROSC register
to include the enabled bit, thus making the returned value one that's
possible on hardware.  Writes to this register are still ignored.

I still haven't compiled this :)

Signed-off-by: Palmer Dabbelt <palmer@sifive.com>